Updated project to version 1.0.2

+++ chromepass/chrome_linux.py
@@ -23,10 +23,15 @@ class ChromeLinux(Chrome):
 
         self.login_db_path = f"/home/{getuser()}/.config/google-chrome/Default/Login Data"
         self.tmp_login_db_path = f"/home/{getuser()}/.config/google-chrome/Default/Login_tmp"
-        shutil.copy2(self.login_db_path, self.tmp_login_db_path)  # making a temp copy since login data db is locked while chrome is running
+
+        if os.path.exists(self.login_db_path):
+            shutil.copy2(self.login_db_path, self.tmp_login_db_path)  # making a temp copy since login data db is locked while chrome is running
+        else:
+            print("It seems that no chrome browser is installed! Exiting...")
+            exit(1)
 
         self.iv = b' ' * 16
-        self.password =  'peanuts'.encode('utf8')
+        self.password = 'peanuts'.encode('utf8')
 
         bus = secretstorage.dbus_init()
         collection = secretstorage.get_default_collection(bus)
@@ -40,7 +45,10 @@ class ChromeLinux(Chrome):
 
     def __del__(self):
         """destructor"""
-        os.remove(self.tmp_login_db_path)
+        try:
+            os.remove(self.tmp_login_db_path)
+        except FileNotFoundError as ex:
+            pass
 
     def decrypt_password(self, encrypted_password):
         """ decrypt the given encrypted password
